memora-speech-processing/requirements.txt
# Core dependencies
openai-whisper>=20231117
torch>=2.0.0
torchaudio>=2.0.0

# Diarization
pyannote.audio>=3.1.0
pyannote.core>=5.0.0

# Audio processing
pydub>=0.25.1
soundfile>=0.12.1
librosa>=0.10.1

# NLP and keyword extraction
transformers>=4.36.0
sentence-transformers>=2.3.0
keybert>=0.8.3
spacy>=3.7.0
yake>=0.4.8

# Summarization
sumy>=0.11.0
nltk>=3.8.1

# LLM integration
anthropic>=0.18.0
openai>=1.12.0

# Utilities
python-dotenv>=1.0.0
numpy>=1.24.0
pandas>=2.0.0

# Huggingface hub (fix compatibility)
huggingface_hub>=0.20.0

# Firebase integration (optional)
firebase-admin>=6.0.0
